make text method latest in the chain
